created-task-and-gh.haf

+++ tpl/tpl-index.php
@@ -2,7 +2,7 @@
 <html lang="en" >
 <head>
   <meta charset="UTF-8">
-  <title>CodePen - Task manager UI</title>
+  <title> Mohammad-deljoo-php</title>
   <link rel="stylesheet" href="<?= BASE_URL ?>assets/css/style.css">
 
 </head>
@@ -22,17 +22,16 @@
       </div>
       <div class="menu">
         <div class="title">FOLDERS</div>
-        <ul>
+        <ul class="folder-list">
+        <li class="<?= isset($_GET['folder_id']) ? '' : 'active' ?>"> <i class="fa fa-folder"></i> ALL </li>
+
           <?php foreach ($folders as $folder):?>
+                        <!-- میگه این لینک با این پارامترارسال کن به پست و گت و... بعد فولدر با این ایدی  بیا و نشون بده اسمشم که از دیتابیس گرفتی نشون بده -->
             <li>
-            <!-- میگه این لینک با این پارامترارسال کن به پست و گت و... بعد فولدر با این ایدی  بیا و نشون بده اسمشم که از دیتابیس گرفتی نشون بده -->
             <a href="?folder_id=<?=$folder->id ?>"><i class="fa fa-folder"></i><?= $folder->name ?></a>
-            <a href="?delete_folder=<?=$folder->id ?>" class="remove" > X </a>
-          <!-- #میگه    -->
+            <a href="?delete_folder=<?=$folder->id ?>" class="remove"  onclick="return confirm('آیا برای دلیت کردن مطمِن هستید؟؟!!')" > X </a>
           </li>
-            <?php endforeach;?>
-          <li class="active"> <i class="fa fa-folder"></i>Current folder</li>
-          
+            <?php endforeach;?>          
         </ul>
         <div>
           <input type="text" id="addFolderInput" style="width: 50%" placeholder="add new folder"/>
@@ -42,7 +41,12 @@
     </div>
     <div class="view">
       <div class="viewHeader">
-        <div class="title">Manage Tasks</div>
+        <div class="title" style="width: 50%;">
+
+        
+          <input type="text" id="addTaskInput" style="width: 100%; margin-left: 3%; line-height: 30px; padding: 3px 10px; border-radius: 3px; border: 2px solid blueviolet;" placeholder="add new tasks">       
+
+        </div>
         <div class="functions">
           <div class="button active">Add New Task</div>
           <div class="button">Completed</div>
@@ -52,29 +56,28 @@
       <div class="content">
         <div class="list">
           <div class="title">Today</div>
+
+
           <ul>
-            <li class="checked"><i class="fa fa-check-square-o"></i><span>Update team page</span>
+          <?php if( sizeof($tasks)): ?>  
+
+          <?php foreach ($tasks as $task):  ?>
+            <li class="<?= $task->is_done ? 'checked' : '' ; ?>">
+            <i class="fa <?= $task->is_done ? 'fa-check-square-o' : 'fa-square-o' ; ?>"></i>
+            <span><?= $task->title ?></span>
               <div class="info">
-                <div class="button green">In progress</div><span>Complete by 25/04/2014</span>
+                <span class='created-at'> Created <?= $task->created_at ?></span>
+                <a href="?delete_task=<?= $task->id ?>" class="remove" onclick="return confirm('آیا برای دلیت کردن مطمِن هستید؟؟!!')"> X </a>
               </div>
             </li>
-            <li><i class="fa fa-square-o"></i><span>Design a new logo</span>
-              <div class="info">
-                <div class="button">Pending</div><span>Complete by 10/04/2014</span>
-              </div>
-            </li>
-            <li><i class="fa fa-square-o"></i><span>Find a front end developer</span>
-              <div class="info"></div>
-            </li>
+            <?php endforeach;  ?>
+            <?php else: ?>
+              <li>No Task Here...</li>
+            <?php endif; ?>
+
           </ul>
-        </div>
-        <div class="list">
-          <div class="title">Tomorrow</div>
-          <ul>
-            <li><i class="fa fa-square-o"></i><span>Find front end developer</span>
-              <div class="info"></div>
-            </li>
-          </ul>
+
+
         </div>
       </div>
     </div>
@@ -95,6 +98,10 @@
           method: "post",
           data: {action:"addFolder",folderName:input.val()},
           success: function(response){
+            if(response == '1')
+            {
+              $('<li><a href="#=<?=$folder->id ?>"><i class="fa fa-folder"></i>'+ input.val() +'</a></li>').appendTo('ul.folder-list');
+            }
             alert(response);
           }
         });
